+++ CHANGELOG.md
@@ -104,3 +104,4 @@
 - no ticket - fix locust's breaking changes
 - MVP-563 - Add time to read
 - MVP-564 - Collect country demographics information from spreadsheets
+- GV2-104 - GP2-121 - Tags and content modules